+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Bucket grid can be set manually by right clicking
 - New graphs automatically set to 0:00 on the hour
 - Fixed a problem with the frame when resetting
+- Changed the overall XP to time left to level
+- Added prediction mode, showing how long until you will level in the graph
 
 ### v1.1.0 – 2025-07-21
 - Added “Tome of XP” history window with Event, Hour, and Day toggleable views